Added comments and function magnus
Added comments and docstrings to most all functions.

Added the experimental function magnus. This function acts like the retry_requests, however, it makes 1.5 times the requests in the first round of threading. This brings some trouble to restarting a "paused" session, but if the first session is completed without pauses it reduces the time slightly.