Fixed error in filter and payload_gen issue
FIxed the filter_all function, so that it starts from where it is intended, i.e. a fix was made on the list comprehension.

Added function ast_impact which reacts when and if an asteroid impacts.

Added function impact_stats which summarises some facts about the particular impact.

Fixed payloads, as these were generated one every run, including if answered y/yes on "continue from last session." This meant that previous payloads were change too, which is not ideal.