Refactoring of filter_responses
Added a counter for the number of unique asteroids which did a CA with the Earth.

Added function impact_Earth to avoid repeating code and for simplifying the overall code.

Removed function filter as it became redundant.

Added function ast_impact which gives some of the information about the impact to the terminal.